Improve lovr.mirror;
Currently lovr.mirror is disabled when the desktop driver is active.
This is good because otherwise lovr.mirror will clear the stuff rendered
by lovr.draw, but also means that it's impossible to layer desktop UI
onto the mirror window when using the desktop driver, using the same
codepath as real VR.

This is an attempt to fix this.  The summary is that lovr.draw and
lovr.mirror will get the same Pass object when the desktop driver is
active.  So lovr.draw will first render the "VR" stuff, and then
lovr.mirror gets a chance to render any UI it wants.

One issue here is that any state from lovr.draw will also get carried
over.  Maybe pass:push('state') can be used, or a new pass:reset()
function would make sense.

Another improvement is things are batched into a single Vulkan submit.
It was really annoying to incur the overhead of 2 submits.